requirements.txt
# Core Framework
fastapi==0.111.0

# Security
python-multipart
pyjwt
pwdlib
pwdlib[argon2]

# Data Validation
pydantic==2.9.2

# Server
uvicorn==0.30.1

# Database
asyncpg==0.29.0

# Testing
pytest==8.3.3
pytest-mock==3.14.0
pytest-asyncio==0.23.7
pytest-html==3.2.0
pytest-cov==4.1.0

# HTTP Requests
httpx==0.27.0
